requirements.txt
altgraph==0.10.2
aniso8601==0.82
appnope==0.1.0
astroid==1.4.4
backports.ssl-match-hostname==3.4.0.2
bcrypt==2.0.0
bdist-mpkg==0.5.0
beautifulsoup4==4.4.1
blinker==1.4
bonjour-py==0.3
brewer2mpl==1.4.1
bs4==0.0.1
cassandra-driver==3.2.2
causality==0.0.3
certifi==2015.9.6.2
cffi==1.5.2
colorama==0.3.6
cql==1.4.0
cqlsh==5.0.3
crypto==1.4.1
csvkit==0.9.1
cycler==0.10.0
Cython==0.23.4
dbf==0.94.3
decorator==4.0.4
Django==1.9.6
django-bootstrap-form==3.2.1
dominate==2.1.17
Flask==0.10.1
Flask-Assets==0.11
Flask-Bootstrap==3.3.5.7
Flask-GoogleAuth==0.4
Flask-Login==0.3.2
Flask-Mail==0.9.1
flask-peewee==0.6.5
Flask-PyMongo==0.4.0
Flask-RESTful==0.2.10
Flask-Sendmail==0.1
Flask-SQLAlchemy==2.1
Flask-User==0.6.8
Flask-WTF==0.12
functools32==3.2.3.post2
futures==3.0.5
geopy==1.11.0
ggplot==0.6.8
gnureadline==6.3.3
gunicorn==18.0
hdbscan==0.6.5
honcho==0.5.0
httplib2==0.9.1
ipykernel==4.1.1
ipython==4.0.0
ipython-genutils==0.1.0
ipywidgets==4.1.1
itsdangerous==0.23
JayDeBeApi==0.2.0
jdcal==1.2
Jinja2==2.7.2
JPype1==0.6.1
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.1.1
jupyter-console==4.0.3
jupyter-core==4.0.6
Keras==1.0.1
lazy-object-proxy==1.2.1
lifelines==0.8.0.1
linecache2==1.0.0
lxml==3.4.4
macholib==1.5.1
MarkupSafe==0.18
matplotlib==1.5.1
mistune==0.7.1
modulegraph==0.10.4
Naked==0.1.31
nbconvert==4.1.0
nbformat==4.0.1
networkx==1.11
nltk==3.1
nose==1.3.7
notebook==4.0.6
numexpr==2.5
numpy==1.11.0
oauth2==1.5.211
openpyxl==2.2.0b1
pandas==0.17.0
passlib==1.6.5
path.py==8.1.2
pathlib==1.0.1
patsy==0.4.1
pdfminer==20140328
pexpect==4.0.1
pickleshare==0.5
Pillow==3.1.1
plotly==1.9.1
Postcodes==0.1
psycopg2==2.6.1
ptyprocess==0.5
py2app==0.7.3
pycparser==2.14
pycrypto==2.6.1
pygeohash==1.0.0
Pygments==2.0.2
pylint==1.5.4
pymc==2.3.6
pymc3==3.0
pymongo==3.1.1
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyOpenSSL==0.13.1
pyparsing==2.1.1
pyScss==1.3.4
pytesseract==0.1.6
python-dateutil==2.5.3
pytz==2016.4
PyYAML==3.11
pyzmq==14.7.0
qr==0.6.0
qtconsole==4.1.0
requests==2.6.0
rpy2==2.3.0
scikit-learn==0.17
scipy==0.17.0
seaborn==0.6.0
selenium==2.53.2
setproctitle==1.1.8
shellescape==3.4.1
simplegeneric==0.8.1
simplejson==3.8.1
six==1.10.0
skflow==0.0.1
sklearn==0.0
SQLAlchemy==1.0.9
statsmodels==0.6.1
tensorflow==0.5.0
terminado==0.5
textblob==0.11.0
Theano==0.8.1
thrift==0.9.3
tornado==4.2.1
traceback2==1.4.0
traitlets==4.0.0
unittest2==1.1.0
virtualenv==15.0.1
visitor==0.1.2
webassets==0.11.1
Werkzeug==0.9.4
wrapt==1.10.6
wtf-peewee==0.2.2
WTForms==2.1
xattr==0.6.4
xgboost==0.4a30
xlrd==0.9.4
xmlutils==1.1
zope.interface==4.1.1
